pyproject.toml
[project]
name = "anki-mcp"
version = "0.1.0"
description = "Headless Anki MCP server for creating Japanese vocabulary flashcards"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastmcp>=2.0",
    "anki",
    "pydantic-settings>=2.0",
    "Pillow>=10.0",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
